VCE Classification/requirements.txt
torch>=2.3.0
torchvision>=0.18.0
timm>=1.0.3
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.12.0
matplotlib>=3.8.0
seaborn>=0.13.0
PyYAML>=6.0.1
Pillow>=10.2.0
tqdm>=4.66.0
